+++ src/core/globalstate.ts
@@ -11,6 +11,7 @@ const persistentKeys = [
     "spyListeners",
     "enforceActions",
     "warnOnUnsafeComputationReads",
+    "disableErrorBoundaries",
     "runId"
 ]
 
@@ -90,6 +91,12 @@ export class MobXGlobals {
      * Warn if computed values are accessed outside a reactive context
      */
     warnOnUnsafeComputationReads = false
+
+    /*
+     * Don't catch and rethrow exceptions. This is useful for inspecting the state of
+     * the stack when an exception occurs while debugging.
+     */
+    disableErrorBoundaries = false
 }
 
 export let globalState: MobXGlobals = new MobXGlobals()
@@ -131,4 +138,20 @@ export function resetGlobalState() {
     for (let key in defaultGlobals)
         if (persistentKeys.indexOf(key) === -1) globalState[key] = defaultGlobals[key]
     globalState.allowStateChanges = !globalState.enforceActions
+}
+
+/**
+ * Don't catch and rethrow exceptions. This is useful for inspecting the state of
+ * the stack when an exception occurs while debugging.
+ */
+export function disableErrorBoundaries() {
+    console.warn("WARNING: Debug feature only. MobX will NOT recover from errors if this is on.")
+    globalState.disableErrorBoundaries = true
+}
+
+/**
+ * Opposite of disableErrorBoundaries
+ */
+export function enableErrorBoundaries() {
+    globalState.disableErrorBoundaries = false
 }